♻️ refactor: group graph agent files
