+++ packages/agent-runtime/src/agents/graph/__tests__/fixtures/goal-loop.graph.json
