+++ packages/agent-runtime/src/agents/graph/index.ts
@@ -1,0 +1,1 @@
+export * from './GraphAgent';
